+++ src/main/java/dev/lambdaurora/aurorasdeco/world/gen/feature/WaySignFeature.java
@@ -21,22 +21,21 @@ import com.mojang.datafixers.util.Pair;
 import com.mojang.serialization.Codec;
 import com.mojang.serialization.DataResult;
 import com.mojang.serialization.codecs.RecordCodecBuilder;
-import dev.lambdaurora.aurorasdeco.block.ExtensionType;
 import dev.lambdaurora.aurorasdeco.block.SignPostBlock;
-import dev.lambdaurora.aurorasdeco.block.WallLanternBlock;
 import dev.lambdaurora.aurorasdeco.block.entity.SignPostBlockEntity;
 import dev.lambdaurora.aurorasdeco.item.SignPostItem;
 import dev.lambdaurora.aurorasdeco.mixin.world.ChunkGeneratorAccessor;
 import dev.lambdaurora.aurorasdeco.registry.AurorasDecoRegistry;
 import dev.lambdaurora.aurorasdeco.registry.AurorasDecoTags;
-import dev.lambdaurora.aurorasdeco.registry.LanternRegistry;
 import dev.lambdaurora.aurorasdeco.world.gen.WorldGenUtils;
+import dev.lambdaurora.auroraslanterns.AurorasLanternsRegistry;
+import dev.lambdaurora.auroraslanterns.ExtensionType;
+import dev.lambdaurora.auroraslanterns.block.WallLanternBlock;
 import it.unimi.dsi.fastutil.objects.Object2ObjectArrayMap;
 import it.unimi.dsi.fastutil.objects.ObjectArraySet;
 import net.minecraft.block.Block;
 import net.minecraft.block.BlockState;
 import net.minecraft.block.FenceBlock;
-import net.minecraft.item.Items;
 import net.minecraft.registry.Holder;
 import net.minecraft.registry.HolderSet;
 import net.minecraft.registry.Registries;
@@ -109,7 +108,7 @@ public class WaySignFeature extends Feature<WaySignFeature.Config> {
 
 		sign.setGenerationSettings(new SignPostBlockEntity.GenerationSettings(config.signPostItem(), facing));
 
-		var lantern = LanternRegistry.fromItem(Items.LANTERN).getDefaultState();
+		var lantern = AurorasLanternsRegistry.WALL_LANTERN_BLOCK.getDefaultState();
 		this.setBlockState(world, pos, lantern.with(WallLanternBlock.FACING, facing).with(WallLanternBlock.EXTENSION, ExtensionType.FENCE));
 
 		pos.set(context.getOrigin());